Ka copy to host fix

+++ src/aa_device_binning_kernel.cu
@@ -37,8 +37,8 @@ namespace astroaccelerate {
     int t_out = ( ( blockIdx.x * BINDIVINT ) + threadIdx.x );
     int t_in = 2 * t_out;
 
-    size_t shift_one = ( (size_t)(c*out_nsamp) + (size_t)t_out );
-    size_t shift_two = ( (size_t)(c*in_nsamp)  + (size_t)t_in );
+    size_t shift_one = ( (size_t)(c)*(size_t)(out_nsamp) + (size_t)t_out );
+    size_t shift_two = ( (size_t)(c)*(size_t)(in_nsamp)  + (size_t)t_in );
 
     d_output[( shift_one )] = (float) ( ( d_input[( shift_two )] + d_input[(size_t)(shift_two + 1)] )/2.0f );
